sorted arrow function by type #114
