+++ index.html
@@ -20,5 +20,5 @@
         </select>
     </center>
 </div>
-<script src="./util.js" type="module"></script>
+<script src="./game.js" type="module"></script>
 </body>
